Account for long long values

+++ 0x0C-more_malloc_free/101-mul.c
@@ -32,6 +32,6 @@ int main(int argc, char **argv)
 			}
 		}
 	}
-	printf("%d\n", (atoi(argv[1]) * atoi(argv[2])));
+	printf("%lld\n", (atoll(argv[1]) * atoll(argv[2])));
 	return (0);
 }
